+++ quickNat_pytorch/log_utils.py
@@ -10,7 +10,7 @@ import math
 import pandas as pd
 import os
 import shutil
-import quickNat_pytorch.evaluator_utils as eu
+import quickNat_pytorch.evaluator as eu
 
 plt.switch_backend('agg')
 plt.axis('scaled')
@@ -29,7 +29,7 @@ class LogWriter(object):
             'train' : SummaryWriter(train_log_path),
             'val' : SummaryWriter(val_log_path)
         }
-        
+        self.curr_iter = 1
         self.cm_cmap = cm_cmap
         self.init_cm()
         self.init_ds()
@@ -55,65 +55,27 @@ class LogWriter(object):
         del cm_batch, batch_output       
             
     def cm_per_epoch(self, phase, epoch, i_batch):
+        print("Confusion Matrix...", end='')
         cm = (self._cm[phase] / (i_batch + 1)).cpu().numpy()
-        plot_cm('confusion_matrix', phase, cm, epoch)
-        self.init_cm()        
-        
-    def update_dice_score_per_iteration(self, predictions, correct_labels, epoch):
-        _, batch_output = torch.max(predictions, dim=1)
-        score_vector = eu.dice_score_perclass(batch_output, correct_labels, self.num_class)
-        self._ds +=  score_vector.cpu()
-        
-    def dice_score_per_epoch(self, epoch, i_batch):
-        ds = (self._ds / (i_batch + 1)).cpu().numpy()
-        plot_dice_score(self,'dice_score_per_epoch', ds, 'Dice Score', epoch)      
-        self.init_ds() 
-        
-    def plot_dice_score(self, caption, ds, title, step=None):
-        tick_marks = np.arange(self.num_class)
-        fig = matplotlib.figure.Figure(figsize=(7, 3), dpi=180, facecolor='w', edgecolor='k')
-        ax = fig.add_subplot(1, 1, 1)            
-        ax.set_xlabel(title, fontsize=10)  
-        ax.xaxis.set_label_position('top')            
-        ax.bar(np.arange(self.num_class), ds)
-        ax.set_xticks(tick_marks)
-        c = ax.set_xticklabels(self.labels, fontsize=8, rotation=-90,  ha='center')
-        ax.xaxis.tick_bottom()
-        if step:
-            self.writer['val'].add_figure(caption, fig, step)
-        else:
-            self.writer['val'].add_figure(caption, fig)
-
-    def image_per_epoch(self, prediction, ground_truth, phase, epoch):
-        ncols = 2
-        nrows = len(prediction)
-        fig, ax = plt.subplots(nrows=nrows, ncols=ncols, figsize=(10, 20))
-        
-        for i in range(nrows):
-            ax[i][0].imshow(prediction[i], cmap = 'jet', vmin=0, vmax=self.num_class-1)
-            ax[i][0].set_title("Predicted", fontsize=10, color = "blue")
-            ax[i][0].axis('off')
-            ax[i][1].imshow(ground_truth[i], cmap = 'jet', vmin=0, vmax=self.num_class-1)
-            ax[i][1].set_title("Ground Truth", fontsize=10, color = "blue")
-            ax[i][1].axis('off')
-        fig.set_tight_layout(True)  
-        self.writer[phase].add_figure('sample_prediction/' + phase, fig, epoch)             
+        self.plot_cm('confusion_matrix', phase, cm, epoch)
+        self.init_cm()
+        print("DONE")
         
         
     def plot_cm(self, caption, phase, cm, step=None):
-        fig = matplotlib.figure.Figure(figsize=(7, 7), dpi=180, facecolor='w', edgecolor='k')
+        fig = matplotlib.figure.Figure(figsize=(8, 8), dpi=180, facecolor='w', edgecolor='k')
         ax = fig.add_subplot(1, 1, 1)
                 
         ax.imshow(cm, interpolation = 'nearest', cmap=self.cm_cmap)
         ax.set_xlabel('Predicted', fontsize=7)        
-        ax.set_xticks(tick_marks)
+        ax.set_xticks(np.arange(self.num_class))
         c = ax.set_xticklabels(self.labels, fontsize=4, rotation=-90,  ha='center')
         ax.xaxis.set_label_position('bottom')
         ax.xaxis.tick_bottom()
 
         ax.set_ylabel('True Label', fontsize=7)
         ax.set_yticks(np.arange(self.num_class))
-        ax.set_yticklabels(classes, fontsize=4, va ='center')
+        ax.set_yticklabels(self.labels, fontsize=4, va ='center')
         ax.yaxis.set_label_position('left')
         ax.yaxis.tick_left()
 
@@ -128,6 +90,49 @@ class LogWriter(object):
         else:
             self.writer[phase].add_figure(caption + '/' + phase, fig)
             
+        
+    def update_dice_score_per_iteration(self, predictions, correct_labels, epoch):
+        _, batch_output = torch.max(predictions, dim=1)
+        score_vector = eu.dice_score_perclass(batch_output, correct_labels, self.num_class)
+        self._ds +=  score_vector.cpu()
+        
+    def dice_score_per_epoch(self, epoch, i_batch):
+        print("Dice Score...", end='')
+        ds = (self._ds / (i_batch + 1)).cpu().numpy()
+        self.plot_dice_score('dice_score_per_epoch', ds, 'Dice Score', epoch)      
+        self.init_ds() 
+        print("DONE")
+        
+    def plot_dice_score(self, caption, ds, title, step=None):
+        fig = matplotlib.figure.Figure(figsize=(7, 4), dpi=180, facecolor='w', edgecolor='k')
+        ax = fig.add_subplot(1, 1, 1)            
+        ax.set_xlabel(title, fontsize=10)  
+        ax.xaxis.set_label_position('top')            
+        ax.bar(np.arange(self.num_class), ds)
+        ax.set_xticks(np.arange(self.num_class))
+        c = ax.set_xticklabels(self.labels, fontsize=8, rotation=-90,  ha='center')
+        ax.xaxis.tick_bottom()
+        if step:
+            self.writer['val'].add_figure(caption, fig, step)
+        else:
+            self.writer['val'].add_figure(caption, fig)
+
+    def image_per_epoch(self, prediction, ground_truth, phase, epoch):
+        print("Sample Images...", end='')
+        ncols = 2
+        nrows = len(prediction)
+        fig, ax = plt.subplots(nrows=nrows, ncols=ncols, figsize=(10, 20))
+        
+        for i in range(nrows):
+            ax[i][0].imshow(prediction[i], cmap = 'jet', vmin=0, vmax=self.num_class-1)
+            ax[i][0].set_title("Predicted", fontsize=10, color = "blue")
+            ax[i][0].axis('off')
+            ax[i][1].imshow(ground_truth[i], cmap = 'jet', vmin=0, vmax=self.num_class-1)
+            ax[i][1].set_title("Ground Truth", fontsize=10, color = "blue")
+            ax[i][1].axis('off')
+        fig.set_tight_layout(True)  
+        self.writer[phase].add_figure('sample_prediction/' + phase, fig, epoch) 
+        print('DONE')
         
     def graph(self, model, X):
         self.writer['train'].add_graph(model, X)
